Various refactoring, cleanup, etc.
Working pretty good. First fadeout isn't working for some reason. CPU/memory usage is too high.

+++ sm-audiokit/Audio/Sample.swift
@@ -10,9 +10,9 @@ import Foundation
 struct Sample {
     var id: String
     var url: URL
-    var duration: Float
+    var duration: Double
 
-    init(id: String, url: URL, duration: Float) {
+    init(id: String, url: URL, duration: Double) {
         self.id = id
         self.url = url
         self.duration = duration
